refactor(telemetry): fix header extraction and reorganize tests
This commit fixes Node.js IncomingMessage header extraction and reorganizes
telemetry tests for better clarity and maintainability.

## Fixed Issues

1. **Node.js header extraction** (otel-core.ts)
   - Fixed span name format: "HTTP GET /test" → "GET /test" (matches Bun.serve)
   - Added missing attributes: http.user_agent, http.host, http.scheme
   - Simplified by using high-level telemetry callbacks

2. **Test organization** (packages/bun-otel/)
   - Split monolithic index.test.ts (348 lines) into focused files:
     • basic.test.ts - BunSDK basic functionality
     • context-propagation.test.ts - W3C trace context
     • node-http.test.ts - Node.js integration
     • resources.test.ts - Resource configuration

3. **Self-documenting tests** (test/js/bun/)
   - Added concise headers explaining what NOT to test
   - Renamed tests to clearly describe what they verify
   - Removed unnecessary mock methods

4. **Removed misplaced test** (telemetry-headers.test.ts)
   - Was testing unsupported behavior (Node.js onResponseHeaders)
   - Node.js uses _node_binding hooks, not high-level callbacks

5. **Documentation** (TEST_STRATEGY.md)
   - Explains core vs bun-otel test separation
   - Provides examples and guidelines for new tests

## Test Results

✅ Core tests: 29 pass, 0 fail (757ms)
✅ bun-otel tests: 8 pass, 0 fail (1.74s)

## Architecture Notes

- **Bun.serve**: Uses high-level callbacks (onRequestStart, onResponseHeaders)
- **Node.js http**: Uses _node_binding hooks (handleIncomingRequest, handleWriteHead)
- **Core tests**: Verify insertion points are called
- **bun-otel tests**: Verify OTel integration and semantic behavior
