--- test/js/bun/http/telemetry-headers.test.ts
@@ -1,134 +1,0 @@
-import { expect, test } from "bun:test";
-import * as http from "node:http";
-import { waitForEvents } from "../telemetry-test-utils";
-
-test("Bun.serve passes headers to telemetry", async () => {
-  const events: Array<{ type: string; id: number; data?: any }> = [];
-
-  // Configure telemetry with headers support
-  Bun.telemetry.configure({
-    onResponseHeaders(id, statusCode, contentLength, headers) {
-      events.push({
-        type: "headers",
-        id,
-        data: {
-          statusCode,
-          contentLength,
-          headers,
-          // Test that we can access headers
-          contentType: headers?.get?.("content-type") ?? headers?.["content-type"],
-          customHeader: headers?.get?.("x-custom-header") ?? headers?.["x-custom-header"],
-        },
-      });
-    },
-  });
-
-  using server = Bun.serve({
-    port: 0,
-    fetch() {
-      return new Response("test body", {
-        status: 201,
-        headers: {
-          "Content-Type": "text/plain",
-          "X-Custom-Header": "test-value",
-        },
-      });
-    },
-  });
-
-  await fetch(`http://localhost:${server.port}/`);
-  await waitForEvents(events, ["headers"]);
-
-  expect(events.length).toBe(1);
-  expect(events[0].data.statusCode).toBe(201);
-  expect(events[0].data.contentType).toBe("text/plain");
-  expect(events[0].data.customHeader).toBe("test-value");
-
-  Bun.telemetry.disable();
-});
-
-test("Node.js http.createServer passes headers to telemetry", async () => {
-  const events: Array<{ type: string; id: number; data?: any }> = [];
-
-  // Configure telemetry with headers support
-  Bun.telemetry.configure({
-    onResponseHeaders(id, statusCode, contentLength, headers) {
-      events.push({
-        type: "headers",
-        id,
-        data: {
-          statusCode,
-          contentLength,
-          headers,
-          // Test that we can access headers (plain object for Node.js)
-          contentType: headers?.["content-type"],
-          customHeader: headers?.["x-custom-header"],
-        },
-      });
-    },
-  });
-
-  const server = http.createServer((req, res) => {
-    res.writeHead(200, {
-      "Content-Type": "application/json",
-      "X-Custom-Header": "node-value",
-    });
-    res.end('{"ok":true}');
-  });
-
-  await new Promise<void>((resolve, reject) => {
-    server.listen(0, () => resolve());
-    server.on("error", reject);
-  });
-
-  const address = server.address();
-  if (!address || typeof address === "string") {
-    throw new Error("Server address not available");
-  }
-
-  await fetch(`http://localhost:${address.port}/test`);
-  await waitForEvents(events, ["headers"]);
-
-  expect(events.length).toBe(1);
-  expect(events[0].data.statusCode).toBe(200);
-  expect(events[0].data.contentType).toBe("application/json");
-  expect(events[0].data.customHeader).toBe("node-value");
-
-  await new Promise<void>(resolve => {
-    server.close(() => resolve());
-  });
-
-  Bun.telemetry.disable();
-});
-
-test("Headers are only accessible during callback", async () => {
-  let savedHeaders: any = null;
-
-  Bun.telemetry.configure({
-    onResponseHeaders(id, statusCode, contentLength, headers) {
-      // Save headers to test later
-      savedHeaders = headers;
-    },
-  });
-
-  using server = Bun.serve({
-    port: 0,
-    fetch() {
-      return new Response("test", {
-        headers: { "X-Test": "value" },
-      });
-    },
-  });
-
-  await fetch(`http://localhost:${server.port}/`);
-  await Bun.sleep(10);
-
-  // Headers should have been captured
-  expect(savedHeaders).toBeDefined();
-
-  // Note: In the current implementation with protect/unprotect,
-  // headers remain accessible after the callback (this is actually OK)
-  // but the documentation should warn that this behavior may change
-
-  Bun.telemetry.disable();
-});
